auth/requirements.txt
flask[async]==2.1.2
flasgger==0.9.5
flask_sqlalchemy==2.5.1
psycopg2-binary==2.9.3
pydantic[dotenv]==1.9.1
gevent==21.12.0
redis==4.3.3
apispec==5.2.2
marshmallow==3.16.0
flask_restful==0.3.9
marshmallow-dataclass==8.5.8
pycryptodomex==3.15.0
icecream==2.1.2
flask-alembic==2.0.1
flask_migrate==2.5.1
Flask-OpenTracing==1.1.0
jaeger-client==4.8.0
user-agents==2.2.0
requests==2.28.1